Moved image download into a seperate service
