+++ frontend/src/app/app.module.ts
@@ -16,7 +16,7 @@ import {HttpClientModule} from '@angular/common/http';
 import {NgxDropzoneModule} from 'ngx-dropzone';
 import {ColorPickerModule} from 'ngx-color-picker';
 import {ColorpickerComponent} from './colorpicker/colorpicker.component';
-import {RestService} from "./rest-upload/rest.service";
+import {RestService} from "./rest-upload/Services/rest.service";
 
 
 @NgModule({
